paring down config handling
vutil.pm and vntrseek.pl seem to be fraught over how the config needs to be juggled and tested. It becomes an illegible mess that's hard to adapt to new parameters. This is the beginning of cutting back on the juggling and making the options loading more salient and to the point.